Created the DTOs and official badge service interface
