+++ backend/FlashIdBackend/Application/Features/Credentials/Exceptions/CredentialExceptions.cs
@@ -23,15 +23,3 @@ public class InvalidDisclosedFieldsException : Exception
     public InvalidDisclosedFieldsException(IEnumerable<string> fields)
         : base($"Invalid or missing disclosed fields: {string.Join(", ", fields)}") { }
 }
-
-public class OfficialNotFoundException : Exception
-{
-    public OfficialNotFoundException()
-        : base("No official record was found for this user.") { }
-}
-
-public class InvalidBadgeTokenException : Exception
-{
-    public InvalidBadgeTokenException()
-        : base("This badge token is invalid, expired, or could not be verified.") { }
-}
