+++ backend/FlashIdBackend/Application/Features/Officials/DTOs/GenerateBadgeTokenResponseDto.cs
@@ -1,0 +1,7 @@
+namespace Application.Features.Officials.DTOs;
+
+public class GenerateBadgeTokenResponseDto
+{
+    public string Token { get; set; } = string.Empty;
+    public DateTime ExpiresAt { get; set; }
+}
