+++ backend/FlashIdBackend/Application/Features/Officials/DTOs/VerifyBadgeRequestDto.cs
@@ -1,0 +1,6 @@
+namespace Application.Features.Officials.DTOs;
+
+public class VerifyBadgeRequestDto
+{
+    public string Token { get; set; } = string.Empty;
+}
